requirements.txt
langchain==0.2.13
ipykernel
python-dotenv
langchain_community==0.2.12
pypdf
pymupdf==1.21.1
bs4
arxiv
certifi
wikipedia
langchain-text-splitters
lxml
langchain-openai>=0.1.0,<0.2.0
langchain-core>=0.2.38,<0.3.0
chromadb
sentence_transformers
langchain-huggingface<0.1.0
langchain-groq==0.1.10
faiss-cpu
streamlit
fastapi
langserve==0.0.41
httpx==0.27.0
sse-starlette
pydantic==1.10.13


